+++ R/FLOSS.R
@@ -14,7 +14,7 @@
 #' @param readLengths Read length used for calculation
 #' @param level Transcript or gene level
 #' @param draw Plot FLOSS vs total reads or not.
-#' @param ignore.seqlevelStyle Ignore the sequence name style detection or not.
+#' @param ignore.seqlevelsStyle Ignore the sequence name style detection or not.
 #' @return A data frame with colnames as id, FLOSS, totalReads,
 #' wilcox.test.pval, cook's distance.
 #' @importFrom IRanges findOverlaps
@@ -49,7 +49,7 @@
 
 FLOSS <- function(reads, ref, CDS, readLengths=c(26:34),
                   level=c("tx", "gene"), draw=FALSE,
-                  ignore.seqlevelStyle=FALSE){
+                  ignore.seqlevelsStyle=FALSE){
   stopifnot(is(reads, "GRanges"))
   stopifnot(length(reads$qwidth)==length(reads))
   level <- match.arg(level)
@@ -59,7 +59,7 @@ FLOSS <- function(reads, ref, CDS, readLengths=c(26:34),
      length(CDS$gene_id)!=length(CDS)){
     stop("CDS must be output of prepareCDS")
   }
-  reads <- fixSeqlevelsStyle(reads, CDS, ignore.seqlevelStyle)
+  reads <- fixSeqlevelsStyle(reads, CDS, ignore.seqlevelsStyle)
   stopifnot(is.numeric(readLengths))
   readLengths <- as.integer(readLengths)
   reads <- reads[reads$qwidth>=min(readLengths) &